MB-44626: flush accept messages after we kill the acceptor
We receive accept messages from the acceptor as a consequence of cb_dist
being the middle man for handling the erlang distribution.

When we remove a particular proto we make sure to kill the acceptor
before we do so we can potentially have message(s) in our mailbox
from the killed acceptor, make sure we clean up the messages
before proceeding.

Change-Id: I961a177df79a7e5657776a9e2dd4e11cd91d65e7
Reviewed-on: http://review.couchbase.org/c/ns_server/+/147393
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>