MB-48047:[BP] flush accept messages after we kill the acceptor
We receive accept messages from the acceptor as a consequence of cb_dist
being the middle man for handling the erlang distribution.

When we remove a particular proto we make sure to kill the acceptor
before we do so we can potentially have message(s) in our mailbox
from the killed acceptor, make sure we clean up the messages
before proceeding.

Backports change related to MB-44626.

Reviewed-on: http://review.couchbase.org/c/ns_server/+/147393

Change-Id: I961a177df79a7e5657776a9e2dd4e11cd91d65e7
Reviewed-on: http://review.couchbase.org/c/ns_server/+/164945
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker