MB-44758 Don't decimate possibly empty timeframes
If the last decimation time was longer ago than the current time minus
twice the pruning interval then the later will be used as the starting
point for the decimation.
This assumes, and is most likely the case, that the last decimation time
not being updated means we weren't gathering stats.

Change-Id: I616d795dd4d5fb072c79d5bb5e7a65e90dc31a07
Reviewed-on: http://review.couchbase.org/c/ns_server/+/147968
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>