MB-44777 Deprecate sasl_password in REST APIs
The sasl_password bucket property if specified in a bucket REST request
will now be ignored and will return a null value in any REST response.

Add "asserts" to enforce assumptions that sasl_password will only be a
bucket property if a cluster is mixed-version. Once the cluster
transitions to cheshire-cat compat-mode this will no longer be the case.

Change-Id: I6167914be2789722f2989dbe97f99912a6bf44d7
Reviewed-on: http://review.couchbase.org/c/ns_server/+/148430
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>