MB-44777 Minimize support for sasl_password
The sasl_password bucket property is no longer used. So the saslPassword
argument will be ignored on bucket create/update and an empty string ""
will be used.

The reason for this is to prevent a password from being stored in
chronicle.

In a subsequent release it will be removed entirely.

Change-Id: I598290a1c6cccc8aa18b9cd84fe8a6597766298b
Reviewed-on: http://review.couchbase.org/c/ns_server/+/149105
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>