MB-44824 partial revert of the commit:
http://review.couchbase.org/c/ns_server/+/148341

Due to this:
1. There's nothing that would protect against RPCs arriving late.
2. For instance, node A is a leader and sends
   maybe_cleanup_old_buckets() to all nodes.
3. One of the RPCs gets stuck somewhere.
4. Rebalance crashes after a timeout.
5. User creates a new bucket.
6. The stuck RPC finally makes to the node. Now the bucket from
   (5) will be considered "old" and its files will get deleted.

It's possible to construct other scenarios of a similar sort.
The old approach is also raceful because the cleanup and bucket
creations are not serialized properly within a node. But with
this change we may be widening the race window. Addressing the
race properly is also quite likely a significant undertaking.

Change-Id: Ifba91c62c7da2e5b5da3259e30d5889c3a611c31
Reviewed-on: http://review.couchbase.org/c/ns_server/+/149949
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>