MB-45204: Set {erl_dist, true} when testing OTP TLS connectivity
in order to use the same environment as net_kernel when connecting
to other nodes

Erlang uses separate caches for connections with erl_dist flag set.
Usage of other caches may affect our ability to connect, that’s the
reason why I want to set this flag when testing the otp connectivity.
Of course, it is wrong that connectivity depends on this flag and it
should not happen when everything works well, but it case of a
problem it should at least break in those places where it’s supposed
to break. Right now I see that the otp connectivity check passes and
we cannot establish real otp connection right after that. I think
this flag is the reason why connectivity checks pass in this case
(at least that's the only difference between those connect attempts).
Setting of this flag will not fix the problem, but at least we should
start seeing the error in the right place - when testing connectivity.

Change-Id: I08411eb66d554638beea10725a711935f29866c1
Reviewed-on: http://review.couchbase.org/c/ns_server/+/150058
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>