MB-45477 [BP] Set initial vbucket state to replica
This is a backport of MB-45376.

At the beginning of a vbucket move during rebalance or graceful failover
the initial vbucket state is set to replica. This effectively reverts
http://review.couchbase.org/c/ns_server/+/57640.

The change being reverted set the initial vbucket state to "pending" to
give the vbucket preferential treatment wrt eviction (vs "replica").
However the change lead to issues where clients (e.g. SDK) accessed the
"pending" vbucket and was blocked until the vbucket transitioned to
"active".

Change-Id: Ie2af8b43c5265c86cc99df8800216e3db55682ca
Reviewed-on: http://review.couchbase.org/c/ns_server/+/150183
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>