MB-44705: Ignore negative histogram values
It might happen in case of time change when os:system_time is used
for measurements. It is also possible to see that in case of
monotonic time usage, though (seen it one time, most likely a bug
in vm specific to mac os).

Change-Id: If7e281ea8aa437f269282256dcbd4dd00c191791
Reviewed-on: http://review.couchbase.org/c/ns_server/+/150327
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>