MB-45487 Use the temp directory provided by the user
Prior to this change, we were using the default temporary directory for
the underlying OS when uploading logs via 'curl'.

We will now use the directory provided by the user and only fallback in
the event that the user has not used the '--temp-dir' flag.

Change-Id: I23d4a8eb997500eeedc9a3a74d80bbf1ea85ec31
Reviewed-on: http://review.couchbase.org/c/ns_server/+/150637
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: James Lee <james.lee@couchbase.com>