Stop using chronicle_rsm:sync/3.
Use chronicle_kv:sync/2 instead. The latter is the function that I
meant to expose externally.

As a side note, the function doesn't take the consistency level
anymore (i.e. quorum or leader). This is because I removed the support
for what used to be 'leader' consistency. So chronicle_kv:sync(Name,
Timeout) behaves as chronicle_kv:sync(Name, quorum, Timeout)
previously would.

Change-Id: Ib72e6882960f52fedfcec366c0f195f8b9256a81
Reviewed-on: http://review.couchbase.org/c/ns_server/+/150683
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>