MB-45107 Use bucket uuid to calculate hash for poolsStreaming/default URI
Today we compute the hash of bucket names and node status when
we generate a reponse for the poolsStreaming/default URI.

This can miss a bucket-addition, bucket-deletion and re-addition
sequence of a bucket with the same name.

In this fix, we change the hash to be computed with the bucket-uuid
instead of the bucket-name.

Change-Id: I3a09d4d0d3e1806c6040050082367cd544089fd7
Reviewed-on: http://review.couchbase.org/c/ns_server/+/151003
Tested-by: <hareen.kancharla@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>