Revert "MB-44679 do remove_peer operation in a reversed order"
This reverts commit 0406be361422be9350b224350851560b2b7a6d29.

Reason for revert:
Imagine that the node that's being removed is a chronicle leader.
Once we remove the node from chronicle topology, it can't make
any changes anymore since it won't learn about the new leader.
Therefore the kv part of the removal will time out.

if we attempt to remove any node from itself (irrespective of
where the chronicle leader is), it's not guaranteed to work.

Also, since mb_master runs off nodes_wanted, if we remove a node
that is the mb_master leader from the chronicle topology, but
fail to remove it from nodes_wanted, the node will continue to
be the mb_master leader, while not being able to write anything
to chronicle.

Change-Id: Ifbb6aa134e9ba23c89fe2beb687524c7b6cc9969
Reviewed-on: http://review.couchbase.org/c/ns_server/+/151289
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>