MB-45646 Don't save last pruning times in the ns_config
We no longer will keep the last_decimation_time and last_truncation_time
in the ns_config. Instead we keep a last pruning time in memory and use
that instead. The last pruning time is set initially to the boot time.
As pruning occurs we note the start time and then kick off the pruning
process. When the process successfully exits the last_decimation_time is
set to the noted start time.

Change-Id: I21a80c1a7d28790b733f4d6d2c3634e5498123d2
Reviewed-on: http://review.couchbase.org/c/ns_server/+/151812
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>