MB-45906: Fix goport args discrepancy between CE/EE versions
The -serverSslPort argument was added recently for the fts service but did not
take into account the differences between CE and EE versions of the product. In
the CE version the ssl_rest_port is not populated and causes a failure to match
when using the community version. Now we optionally include it if we are
currently on the EE version, and not in CE versions.

Change-Id: I3aa85cee6b48f6a72971f022aaca32de6ccbb0c7
Reviewed-on: http://review.couchbase.org/c/ns_server/+/152121
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>