MB-45703 Log node when service_agents die
Prior to this change when a rebalance was happening and a service agent
died it was difficult to determine on which node the failure occurred.
This change adds the node to the failure result so it gets logged on
the node doing the rebalance.

Change-Id: I8d19abcb441f54f7c313f2df84f3086748df6fc6
Reviewed-on: http://review.couchbase.org/c/ns_server/+/152367
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>