MB-46010 Authenticate cb-on-behalf-of usr using special usr creds
Services (like Analytics) need to foward pluggable UI requests to
remote nodes running the same Service. We are advocating that they
add a "cb-on-behalf-of" HTTP header, so that the remote nodes can
authenticate the request.

1) With this patch we are adding new logic to menelaus_auth layer to
authenticate the user encoded in the "cb-on-behalf-of" header,
using the special user ("@") and the admin creds encoded in the
Basic Auth Header.

2) When the compat mode is < 7.0.0, we send the local node info too
in the token. This is useful in the case when the Service running on
the local node, decides to forward the request to other remote Nodes.

Change-Id: I8db5bcc74fc4bc0d38494741e6e8fb3f5f5152be
Reviewed-on: http://review.couchbase.org/c/ns_server/+/152624
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>