MB-45853 Don't pass float to timer:apply_after
When renaming of a file in rename_and_refresh encountered an error we
retry the rename after a delay from timer:apply_after but in the process
of setting up the retry the sleep time was converted to a float.
timer:apply_after expects an integer.

Change-Id: I5538a6ed92a448bc345b466a4cacb6d79b91b96c
Reviewed-on: http://review.couchbase.org/c/ns_server/+/152815
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>