MB-46040: Pass self-generated node key in engageCluster...
... to joinee so that cluster node can establish a connection to it

When a node is added to a cluster under node-to-node encryption with
self-generated certs, currently the node-being-added and the cluster
node sometimes have a hard time getting a connection established.
During the completeJoin phase both nodes try to connect to each
other, but since the cluster node doesn't have the CA of
the node-being-added, the TLS connection from the cluster node can't
be established (fails with reason Unknow CA). So our only hope is
that the connect from node-being-added will succeed. At the same time
erlang's handshake algorithm is seeing connections in both directions
are being established, so it chooses only the one with greatest node
name and closes the other one. Depending on node names the connection
that is closed happens to be the one from the node-being-added. In
this case completeJoin fails as nodes can't connect to each other.
So basically erlang assumes that both nodes can connect to each other,
which is not true in our case.

Change-Id: Ib37744e96f18da55d471d973ee088fb7e6981479
Reviewed-on: http://review.couchbase.org/c/ns_server/+/152971
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>