Don't print "~n" in a logged warning
In a log message there as a "~n" specified but an empty [] parameter
list. In the guts of ale there's special handling when the list of
parameters is empty which leads to io_lib:format() getting bypassed.

Specifically:

?log_warning("send heartbeat timed out~n", [])

would result in

send heartbeat timed out~n

Change-Id: I5b0708d7f20a7e407cb37228b0494dda83949851
Reviewed-on: http://review.couchbase.org/c/ns_server/+/152848
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: <hareen.kancharla@couchbase.com>