Make sure that node's vclock timestamps are non-decreasing.
Conflict resolution uses these timestamps to resolve conflicts. The
function that resolves conflicts needs to be associative (among other
things). With the possibility of node timestamps going back, it's not
associative.

Consider this scenario.

 1. Node A makes a change. A's time is 100.
 2. Node B makes a change without seeing the change from A. B's time
    is 99.
 3. Node A makes another change, but in the meanwhile, its time went
    back. A still hasn't seen the change made by B. A's time is 95.

Node C observes these events. If it observes (1), (2), and then (3),
it will pick the value from (1). If it just observes (2) and
then (3) (which is valid, because (3) is a descendant of (1)), it will
pick the value from (2).

Note, that in order for the merge procedure to be convergent, the way
the ties are broken when the timestamps are equal is crucial. I don't
believe that the way it's currently done is strictly convergent. But
changing this behavior is backward incompatible and may actually lead
to non-convergence in mixed clusters. So I decided not to touch it
now.

Change-Id: I19d2f2041b70eb5d9718d3be50a3c6dd3ee302e5
Reviewed-on: http://review.couchbase.org/c/ns_server/+/153077
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>