MB-46105: Set appropriate defaults for listeners
On upgrade, we won't find local_listerners in the dist_cfg file as we
never recorded them. Since we no longer start both inet and inet6 listeners
for both local and external, always have the preferred proto as the default
list of listeners.
This avoids the situation where we start the wrong listener on upgrade.

Change-Id: Ieef1c28ac579cbede8d0a8d3d5dd572b71acf843
Reviewed-on: http://review.couchbase.org/c/ns_server/+/153079
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>