Revert "MB-46085 Don't treat deletions specially wrt conflict resolution."
Reverting this commit because ns_server relies on the special
treatment of deletions.

When a node is removed from the cluster and shortly thereafter added
back, if the time on the node orchestrating the removal is
sufficiently ahead of the time on the node that got removed, any
conflict resolutions between these two nodes will favor the node that
orchestrated the removal. But we don't want that. We actually want the
values from the other node to win when it's added back. The special
handling of deletions achieves that more or less: non-deleted values
from the node upon readdition will win the deleted values from the
cluster. While this is not a principled way to deal with this
requirement, we've bean relying on it for a long time and there's
probably no time to address it differently at the moment.

Change-Id: Idf8b7250779c12e9520c5bee1965648e47515f38
Reviewed-on: http://review.couchbase.org/c/ns_server/+/153276
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>