MB-45793 Extend log_{debug, warn, error, critical} to compact large terms
log_{debug, warn, error, critical} now takes a 3rd argument 'Opts',
which can be used to explicitly toggle compaction of any large terms
that might be printed.

Compaction in this context implies, adding an ellipsis ('...') to an
Arg, if chars needed to print a logline exceed chars_limit.

For the linked ticket, we use 'chars_limit' option to reduce the size
of 'Reason' printed in the exit exception.

Change-Id: I1c7722565052963b1b01fb3bdd1eb8f46b873763
Reviewed-on: http://review.couchbase.org/c/ns_server/+/153150
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Meni Hillel <meni.hillel@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>