MB-46233 Always set_initial some keys pre cluster join.
The old code relied on the fact that those keys are already present in
the config. But with tombstone purging this is not guaranteed to be
the case. Without this, those keys that were purged would take values
from the default config and would have vector clocks attached to
them. This in turn would lead to conflicts on join causing the
anomaly observed in the ticket.

Change-Id: Ie87c6914c7f7881076441c2a8dce7d8b4c666ecc
Reviewed-on: http://review.couchbase.org/c/ns_server/+/153492
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>