Make master takeover quicker after old master has been removed
Maintain a list of higher_priority_nodes, i.e., list of nodes
that have higher priority than us to become the master and their
last_seen time.

When nodes_wanted change check if the old master(which is not us) was
ejected from the cluster, and check if there are no other
higher_priority_nodes in the cluster that we have heard from for
10sec(Timeout for master takeover), if so immiediately take over
mastership.

Care is taken that a newly initialized node doesn't become master
immediately before having the chance to hear from other nodes in the
cluster. We do this by initialize higher_priority_nodes with
{node(), <Initialization time>} tuple.

Change-Id: I3bbf94c62a5ef48b7e8dd3f904766cdfd07c9e49
Reviewed-on: http://review.couchbase.org/c/ns_server/+/153851
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>