MB-46143: also format log lines with empty arguments
Previously, if you logged something with ale, but did not include any
arguments that needed to be interpolated, ale would skip the formatting
step entirely. This means that we cannot include other formatting
identifiers and expect them to be formatted correctly. This commit
fixes this oversight.

Change-Id: I9a25b4654db16fcbdac643def15a9a3b118d05c7
Reviewed-on: http://review.couchbase.org/c/ns_server/+/154184
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Meni Hillel <meni.hillel@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>