MB-44878: Set correct range vector interval for old stats
In case when stats decimation is enabled, we might not have enough
samples to calculate rates for some zoom settings with standard
range vector interval, so we utilize the step parameter as a marker
that higher range interval should be used.

We assume here that decimation levels are set with honor to zoom
levels. Otherwise all the old stats (not only rates) will be broken
anyway.

Change-Id: I98c5cd5eb8a477ea032c28ca2ddde35c8400e7bf
Reviewed-on: http://review.couchbase.org/c/ns_server/+/156824
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>