MB-44119 Add bucketname and uuid to pools/default
pools/default returns a hash computed from the bucket names and uuids.
Services use this hash to know any changes to the buckets config.
When the hash changes, they make a seperate request to pools/default/buckets
to get the bucket names.

With this change we are including the bucketname and uuid as a part of
the pools/default response to avoid the extra request to
pools/default/buckets.

Change-Id: I362e806ca30b7344ba38be7e80532715f5c24803
Reviewed-on: http://review.couchbase.org/c/ns_server/+/155341
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>