MB-45608 Mask Basic Auth Headers in process dumps
We dump the process info of all the processes on a node, when diag
response generation times-out.
One of those process happened to be lhttpc_client used to forward
pluggable UI requests to Services - the backtrace for that process
leaked the memcached admin user/pwd. This patch prevent dumping that
data.

Change-Id: I847066c0b01c3ee1424e81a5ba8a831039fe6d99
Reviewed-on: http://review.couchbase.org/c/ns_server/+/155502
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: <hareen.kancharla@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>