MB-41763 Remove top_keys per bucket code
top_keys returned via memcached are the "top used" keys per bucket. With
the introduction of collections and scopes, these stats become
meaningless. 'hit counts' to two key (with the same name) in different
collections are accrued against the same entry for the key in the stats.

Change-Id: Ia218d3540cd660ca1060d1580ab3b3d7324c7d7c
Reviewed-on: http://review.couchbase.org/c/ns_server/+/155865
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: <hareen.kancharla@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>