MB-47110: Upgrade Logs Collect Info page to Angular.
"Start Collecting" button is disabled now until the form is valid
for starting collecting logs.
"Get Cluster Summary" modal is closing after the text is copied
with a success/error message.
A new plugin is added for copying text in Angular code: "ngx-clipboard".
A new Angular component is added for the selectable list of nodes:
"mn-selectable-nodes".
A new implementation for "mnStripPortHTML" filter is added in both
Angular and AngularJS.

Change-Id: I513772664baeb5a81d3ab6acdb31d7647fc0d895
Reviewed-on: http://review.couchbase.org/c/ns_server/+/156485
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>