MB-41101 No longer require developer preview for magma
Support for magma no longer requires developer preview. The minimal
release is Neo even though magma was available in 7.0 with developer
preview enabled. The reason is mixed version cluster and upgrades are
not supported when developer preview is enabled (and it cannot be
disabled).

Change-Id: I7e458850aa47928928904c308370e9da57b9f893
Reviewed-on: http://review.couchbase.org/c/ns_server/+/156608
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>