MB-41637 Use cb-on-behalf-of header in pluggable UI requests.
ns_server proxies pluggable UI requests to the respective Services. We
previously sent the “ui-auth-token” HTTP header as a part of the request
that was forwarded to the service. The Service in-turn verified the “ui-
auth-token” by calling the local ns_server (via cbauth), which would in-
turn call the ns_server proxying the original-request to authenticate the
data access.

With this change, we add the on-behalf-of header so that the service can
quickly authenticate the user, avoiding the round trip.

The Basic Auth Header encodes the "@" super user and the memcached admin
password and the "cb-on-behalf-of" Header encodes the user:domain on
whose behalf the authentication is being performed.

Change-Id: Icc7a0548042734c91e56e75d9f0a81326717dee6
Reviewed-on: http://review.couchbase.org/c/ns_server/+/150041
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: <hareen.kancharla@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>