MB-44361: Temporarily save separate cert file for erlang
Erlang of version 22 and lower doesn't support having cert chain in
certfile option. It expects certfile to have 1 cert only, and the rest
of the chain is expected to be in cacert opt. This is true for server
side only.

This change can be reverted when we migrate to erl 23 or higher.

Change-Id: Id0776be9d0506efb2a9a0d60e5ca0d249e722283
Reviewed-on: http://review.couchbase.org/c/ns_server/+/158785
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>