MB-47702: Conditionally kill epmd on settings change and restart
In order to support EnforceTLS(don't listen on unencrypted ports) and
address family only feature, it was decided to kill epmd.
Both the above features are only supported above 7.0.

The reason for killing epmd is that it no longer required from 6.5 and
it is hard to make it listen only on a particular address family. The
ports opened by epmd are also unencrypted ports.

In order, to kill epmd we need to start it with relaxed_command_check
option.
Using "epmd -kill" since we can only kill epmd if,
1. no node names are registered
2. started with relaxed_command_check option

Corresponding Windows change,
http://review.couchbase.org/c/voltron/+/159157

Change-Id: I537e68a14b0455de5e4dd15bf40fcc91e76ad742
Reviewed-on: http://review.couchbase.org/c/ns_server/+/158666
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>