MB-47387 Increase magma bucket creation timeout
For a magma bucket there are 100s of files that must be opened and read.
Doing this is taking longer than the 2 minute timeout. So the default
for magma bucket creation is now 5 minutes.

This can be changed such as in this example to change to 700 seconds:

ns_config:set({node, node(),
              {timeout, {ns_memcached, magma_creation}}}, 700000).

Change-Id: I5b1438a4fcb984c50ed790ac3397a3cd55b90030
Reviewed-on: http://review.couchbase.org/c/ns_server/+/158683
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>