MB-38457: Teach cb_dist to work with encrypted private keys
We can't put passwords to disk together with other ssl_dist_opts,
so we extract and add the password dynamically as soon as we
are able to extract it.

Note that it won't work with vanilla erlang, because ssl_dist_opts
ets is protected there.

Change-Id: I1d4e8cc01ba87d8e71bfb50060c8c3a34c54c5a4
Reviewed-on: http://review.couchbase.org/c/ns_server/+/160243
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>