MB-47942 Remove node scoped autocompaction
The compaction_daemon was doing a node-scoped read of ns_config to get
the autocompaction settings ({node, node(), autocompaction}) but
wasn't listening for changes to that key. Also there's no UI support for
changing the key. Because of these two factors we'll only look for the
global autocompaction settings and never the node-scoped.

Change-Id: Ia3e24137379407fafc45f89eaa2e98acd4872695
Reviewed-on: http://review.couchbase.org/c/ns_server/+/162286
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>