MB-40229 - use ALTER INDEX to delete invidual replicated indexes
If you use DROP INDEX with a replicated index, it drops every
replica. To only drop a single replica requires ALTER INDEX.

In the UI, we can tell if an index is replicated with the "numReplica"
field, but the replicaId field can only be found embedded in the full
index label. Thus we must parse the label to pull out the replicaId.

MB-48460 - don't use 'default:' prefix in mixed clusters

Change-Id: Ief2dd7f15ab62f89d198fe4b348ba6987c68865a
Reviewed-on: http://review.couchbase.org/c/ns_server/+/162515
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>