MB-48522 do not block user updates during the select
As far as I rememberlLocking was done out of abundance of caution,
because selects were done against dets, which was perceived as
unreliable back then.

The consequence of removing the lock is that if update happens during
select it is not determined if it will appear in select results.
It might or it might not depending of the position of cursor.
But I think we can live with that

Change-Id: I1c2183fdcaf789a1a19c19ced39a3a1da30351b2
Reviewed-on: http://review.couchbase.org/c/ns_server/+/163602
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>