MB-43219 Fix bug when magmaFragmentationPercentage isn't specified
A recent change introduced this code:

   Percent =
       compaction_daemon:global_magma_frag_percent(),

where "Percent" is an argument to the function that contains this code.
So this code is trying to do a "match" instead of an assignment. The
code works when the value "Percent" matches the global value. This is
the case by default when creating magma buckets via the UI. This isn't
the case when doing it via something such as:

$ cluster_connect -n 1 -s 1024 -I 512 -M plasma -S magma -T n0:kv

Change-Id: I64c42104fdfd1532ea6cd55df33aa627fc0c9670
Reviewed-on: http://review.couchbase.org/c/ns_server/+/163148
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>