MB-49017: Don't try to migrate certs if cert_and_pkey is missing
basically to support the case when pre-neo package is uninstalled,
and then neo package is installed. In this case we will have old
certs on disk (they are not removed by uninstall procedure), but
the config will be missing and no upgrade should be performed.

Change-Id: I1f9c1c5082bd2245081ef198631ba4dca0f0df54
Reviewed-on: http://review.couchbase.org/c/ns_server/+/164024
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>