MB-48369: Make it possible to pass nodeInit args in clusterInit
... in order to make it possible to init cluster with a single
http call, so cluster initialization doesn't break in the middle
of the process when address family settings are modified, or
when hostname is changed while configuration is done via https

In particular the following scenarios do not work. They leave the
cluster in not initialized state.

Scenario1:
1. User uses raw IPv4 in browser address bar when initializing
   the cluster in the UI;
2. User chooses to use IPv6-only option
3. After IPv6-only option has been applied, UI is unable to continue
   initialization as the old address doesn't work

Scenario2:
1. User uses https when initializing cluster in UI
2. User changes hostname of the cluster
3. After the hostname has been changed, the UI is unable to continue
   initialization because server's certs has changed

This change allows UI to finish the initialization even if the address
that UI uses becomes unavailable after new settings are applied.

Change-Id: Iff15560e7a8fb7551b029022df10e2b6406bd263
Reviewed-on: http://review.couchbase.org/c/ns_server/+/164318
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>