Don't take into account whether nodes are alive.
When showing "Additional active servers required to provide the
desired number of replicas" warning.

There may have been some reasoning behind doing it this way, but
I (and a couple of folks I asked) find this confusing.

Consider a bucket with 2 replicas in a cluster of 3 nodes. If one of
the nodes is down, we'll show the warning. But in a cluster of 4 nodes
where one node is down, the warning won't be shown, even though some
vbuckets don't have the desired number of replicas.

This patch changes the logic to consider all active kv nodes whether
they are currently alive or not.

Change-Id: I0680e32e0b2bbbc314520a2b936527f6814c016a
Reviewed-on: http://review.couchbase.org/c/ns_server/+/164506
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>