Change how we decide whether to generate rack-aware vbucket maps.
The decision is now based on the rack configuration
post-rebalance. This is to handle the situation where the current
state of server groups suggests we should be rack-aware, but
post-rebalance there's only one server group left.

The side-effect of the old behavior was that if we transitioned from
rack-aware to non-rack-aware configuration by rebalancing some nodes
out, the rebalance button would still be enabled.

This patch also changes how we decide whether we are in rack-aware
mode in general. Previously we looked at all non-empty server groups
and if there were more than two, we'd conclude that we needed to be
rack-aware. This is changed to only consider server groups that have
kv nodes. This is done for two reasons:

 1. This made fixing the bug easier.
 2. ns_server is only concerned with kv rack-awareness and the old
 behavior was simply a consequence of services not being at play when
 the code was written.

Change-Id: I8df465f5420a2cc34ff22a4d847af60285927c4a
Reviewed-on: http://review.couchbase.org/c/ns_server/+/164508
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>