MB-44513: Trim paths before setting them
The paths are trimmed anyway in couchdb. If the same API called
one more time and we don't trim them, we will not be able to
figure out that the path has not changed.

Basically if path contains spaces the logic that compares old and
new paths breaks.

Change-Id: Ie5f803dc80cba64c1c53c97f73299bfb7f2d4961
Reviewed-on: http://review.couchbase.org/c/ns_server/+/164681
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>