MB-49376: Alternative external service address validation
Specifically, if you attempt to set an external alternative address
for a service that does not exist on a provisioned node, this is now
disallowed. We will continue to allow setting these values on
non-provisioned nodes, but after initialization, only the services for
which the node actually has are returned (this is how it works now).

Change-Id: I6c35c6dff71338fb75226a17eae1d0190f3f214d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/165419
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>