MB-49451 Add nodeServicesUri to pools/default response
In addition to adding nodeServicesUri to the pools/default response it
will contain a hash of the nodesExt portion of the nodeServices information.
The reason for this is to allow clients to receive notifications from a
single endpoint which can be used to determine whether or not to make
additional calls....vs receiving notifications from multiple endpoints
with partial views of what the caller considers the entire relevant
state.

   "nodeServicesUri": "/pools/default/nodeServices?v=45350433",

Change-Id: I345de2ee366b913fd27aca8852d81a8aecbcd70d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/165565
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>