MB-49450 Add hash of nodes info
In order to assist callers of REST Apis such as  /poolsStreaming/default
and /pools/default a hash of the node portion of the response is added.
This allows callers to check last vs current hash values instead of
checking all the fields in the node portion to determine if something in
that section has changed.

Change-Id: I92ae655ea4c06af58c0efe513484ef99c6995643
Reviewed-on: https://review.couchbase.org/c/ns_server/+/165567
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>