MB-34703 Add Framing info extras in mc_binary to impersonate user
1) Memcached Binary Protocol now has additional "flexible framing
extras" a.k.a "Frame Info objects" which allow us to pass additional
information (in addition to the current extras).

The specific protocol extensions are described/detailed here:
http://src.couchbase.org/source/xref/trunk/kv_engine/docs/
BinaryProtocol.md#61-144

2) When a 'user-x' (via UI/REST) performs a CRUD operations on any
document, we add that user as a 'frame info object' with the
objid = 'impersonate user' as described in the link below.

http://src.couchbase.org/source/xref/trunk/kv_engine/docs/
BinaryProtocol.md#164-174

From memcached POV the operation is now seen as being performed by
"@ns_server" impersonating (or on-behalf-of) the 'user-x'.

Change-Id: Id259b573e7e5d5d08dc4479acf4448f62c8a4c28
Reviewed-on: https://review.couchbase.org/c/ns_server/+/166017
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>