MB-49697: Missing 'rest' in alternate ports update
Because 'rest' is always on the node we need to make sure to add it to
the list of services ran on the node when validating the input. This
regression was created when we started validating the input for
whether or not the service is on the node, except when the node is not
provisioned, in which case we continue to allow the old behavior.

Change-Id: I4acb38e5a7475094df96f4fa29fa20b7e8c47871
Reviewed-on: https://review.couchbase.org/c/ns_server/+/166168
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>