MB-48063: Do quorum read on manifest for collection update
Prior to this patch, we did a local read of the collections
manifest before updating the manifest. On clusters of more than 2
nodes this can result in a manifest update failure when scopes
and collections are created and dropped in quick succession by
an SDK client. (The SDK round  robins requests between nodes and
it can happen that a node that's handling a given request isn't
in possession of the latest manifest, causing the update to
fail.)

Considering performance, in the best case the node performing an
update happens to be the chronicle leader and there is little
overhead. In the worst case, there's an extra round trip from the
node handling the update to the leader + a wait while the latest
chronicle changes are synced.

In some unscientific testing I did on a 4 node cluster on my
laptop, I saw the following times in milliseconds when creating
collections in quick succession, these observations are borne
out (approx 50 collections created):

  Consistency  Min  Max  95th Percentile
  local        360  683  637
  quorum       377  876  668

Ultimately, I feel the performance tax is warranted as the it's
trivial to reproduce the collections update failure in a 3 node
cluster using the Java SDK and this seems like something that
may happen commonly in tests.

Change-Id: I1c179aba8f82185c011537738e37f12d327bb8ce
Reviewed-on: https://review.couchbase.org/c/ns_server/+/166278
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>